+++ packages/main/src/components/AnalyticalTable/Resizer/index.tsx
@@ -1,24 +1,56 @@
-import { useMemo, useCallback } from 'react';
-import React from 'react';
+import React, { RefObject, useCallback, useRef } from 'react';
 
-const Resizer = ({ children, className, ...rest }) => {
-  const classNames = useMemo(() => {
-    let name = 'rt-resizer';
-    if (className) {
-      name += ` ${className}`;
-    }
-    return name;
-  }, [className]);
+const Resizer = (props) => {
+  const { column } = props;
+  const { key } = column.getHeaderProps();
 
-  const handleOnClick = useCallback((e) => {
-    e.preventDefault();
-    e.stopPropagation();
-  }, []);
+  const parentWidth = useRef(0);
+  const startX = useRef(0);
+  const resizerRef: RefObject<HTMLDivElement> = useRef();
+
+  const onResize = useCallback(
+    (e) => {
+      e.stopPropagation();
+      const parentElement = resizerRef.current.parentElement;
+      const pageX = e.pageX;
+      const newWidth = Math.max(parentWidth.current + pageX - startX.current, column.minWidth);
+      parentElement.style.width = `${newWidth}px`;
+    },
+    [column, parentWidth, key, startX, resizerRef]
+  );
+
+  const onEndResize = useCallback(
+    (e) => {
+      e.stopPropagation();
+      document.removeEventListener('mousemove', onResize);
+      document.removeEventListener('mouseup', onEndResize);
+      document.removeEventListener('mouseleave', onEndResize);
+
+      delete resizerRef.current.parentElement.style.userSelect;
+    },
+    [onResize, resizerRef]
+  );
+
+  const onStartResize = useCallback(
+    (e) => {
+      e.stopPropagation();
+      parentWidth.current = resizerRef.current.parentElement.getBoundingClientRect().width;
+      startX.current = e.pageX;
+      resizerRef.current.parentElement.style.userSelect = 'none';
+
+      document.addEventListener('mousemove', onResize);
+      document.addEventListener('mouseup', onEndResize);
+      document.addEventListener('mouseleave', onEndResize);
+    },
+    [onResize, onEndResize, parentWidth, startX, resizerRef]
+  );
 
   return (
-    <div className={classNames} {...rest} onClick={handleOnClick}>
-      {children}
-    </div>
+    <div
+      style={{ width: '16px', float: 'right', height: '100%', cursor: 'ew-resize' }}
+      onMouseDown={onStartResize}
+      ref={resizerRef}
+    />
   );
 };